requirements.txt
requests
python-dotenv
Flask
gunicorn
pyktok
discord.py